requirements.txt
cryptography>=41.0.0
Pillow>=10.0.0

